+++ plugins/bundler/package.js
@@ -1,7 +1,7 @@
 /* eslint-disable prefer-arrow-callback */
 Package.describe({
-    name: 'skadmin:meteor-desktop-bundler',
-    version: '2.3.0',
+    name: 'communitypackages:meteor-desktop-bundler',
+    version: '3.0.1',
     summary: 'Bundles .desktop dir into desktop.asar.',
     git: 'https://github.com/sharekey/meteor-desktop',
     documentation: 'README.md'
@@ -11,7 +11,7 @@ Package.registerBuildPlugin({
     name: 'meteor-desktop-bundler',
     use: ['ecmascript@0.4.4'],
     sources: ['bundler.js'],
-    npmDependencies: { chokidar: '2.0.3' }
+    npmDependencies: { chokidar: '3.5.3' }
 });
 
 Package.onUse(function onUse(api) {