+++ plugins/watcher/package.js
@@ -1,7 +1,7 @@
 /* eslint-disable prefer-arrow-callback */
 Package.describe({
-    name: 'skadmin:meteor-desktop-watcher',
-    version: '2.3.0',
+    name: 'communitypackages:meteor-desktop-watcher',
+    version: '3.0.1',
     summary: 'Watches .desktop dir and triggers rebuilds on file change.',
     git: 'https://github.com/sharekey/meteor-desktop',
     documentation: 'README.md',
@@ -9,14 +9,14 @@ Package.describe({
 });
 
 Npm.depends({
-    chokidar: '2.0.2'
+    chokidar: '3.5.3'
 });
 
 Package.onUse(function onUse(api) {
     api.versionsFrom('METEOR@1.3.3');
     api.use('ecmascript');
     api.use([
-        'skadmin:meteor-desktop-bundler@2.3.0',
+        'communitypackages:meteor-desktop-bundler@3.0.0',
     ], ['server'], {
         weak: true
     });